+++ src/test/java/gg/top/webhooks/springboot/CustomServer.java
@@ -1,0 +1,11 @@
+package gg.top.webhooks.springboot;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CustomServer {
+  public static void main(final String[] args) throws Exception {
+    SpringApplication.run(CustomServer.class, args);
+  }
+}
