+++ src/test/java/gg/top/webhooks/springboot/CustomServerSecurityConfiguration.java
@@ -1,0 +1,16 @@
+package gg.top.webhooks.springboot;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.web.SecurityFilterChain;
+
+@Configuration
+public class CustomServerSecurityConfiguration {
+  @Bean
+  public SecurityFilterChain filterChain(final HttpSecurity http) {
+    return http.csrf(csrf -> csrf.disable())
+        .authorizeHttpRequests(auth -> auth.anyRequest().permitAll())
+        .build();
+  }
+}
