--- src/test/java/org/discordbots/api/interceptors/PostCommandsInterceptor.java
@@ -1,20 +1,0 @@
-package org.discordbots.api.interceptors;
-
-import okhttp3.HttpUrl;
-
-public class PostCommandsInterceptor extends BaseInterceptor {
-  @Override
-  protected boolean isCorrect(final String method, final String path, final HttpUrl url) {
-    return method.equals("POST") && path.endsWith("/projects/@me/commands");
-  }
-
-  @Override
-  protected int getStatusCode() {
-    return 204;
-  }
-
-  @Override
-  protected String getMessage() {
-    return "No Content";
-  }
-}
