--- src/test/java/org/discordbots/webhooks/springboot/CustomServer.java
@@ -1,11 +1,0 @@
-package org.discordbots.webhooks.springboot;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CustomServer {
-  public static void main(final String[] args) throws Exception {
-    SpringApplication.run(CustomServer.class, args);
-  }
-}
