+++ src/test/resources/GetSelfResponse.json
