+++ src/test/resources/GetVoteResponse.json
