+++ src/test/resources/GetVotesResponse.json
