+++ src/test/resources/IntegrationDeletePayload.json
