+++ src/test/resources/VoteCreatePayload.json
